+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,15 @@
 # home
 
+## 0.2.84
+
+### Patch Changes
+
+- Updated dependencies
+  - @versa/examples@2.1.0
+  - @versa/schema@2.1.0
+  - @versa/react@1.14.0
+  - @versa/belt@1.14.0
+
 ## 0.2.83
 
 ### Patch Changes