+++ packages/belt/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @versa/belt
 
+## 1.14.0
+
+### Minor Changes
+
+- basic support for schema version 2.1.0; including service itemizations
+
+### Patch Changes
+
+- Updated dependencies
+  - @versa/schema@2.1.0
+
 ## 1.13.4
 
 ### Patch Changes